Refactors create and edit reminder
Combines create and edit reminder functionality into one file. This
involves breaking some of the logic out into the component in order to
handle the scenarios of editing an old reminder versus adding a new reminder. This commit also revises existing tests to make them pass, as the refactoring had introduced new class names which had previously made many of the tests fail.

name, date, and title show up again in editing mode

edit and save functionality is starting to work

the edit prop is being passed down successfully

issue is done except for rerouting and tests

Reminder form refactor now works

first AddNewReminder acceptance test now passes again

all add new reminder acceptance tests now pass

all acceptance tests now pass

all tests now pass again

completed code cleanup to do PR
