--- app/templates/components/new-reminder-form.hbs
@@ -1,17 +1,0 @@
-<form class="create-new--form" {{action 'createReminder' on="submit"}}>
-  <label>
-    Title
-    {{input class="add-title-input" value=title placeholder="Title..."}}
-  </label>
-  <label>
-    Date
-    {{input class="add-date-input" value=date placeholder="Date..."}}
-  </label>
-  <label>
-    Notes
-    {{textarea class="add-notes-input" value=notes placeholder="Notes..."}}
-  </label>
-  {{input class="input-submit" type="submit"}}
-</form>
-
-{{outlet}}
