+++ app/templates/reminders/new.hbs
@@ -1,3 +1,3 @@
 <h2>Create a New Reminder</h2>
 
-{{new-reminder-form}}
+{{reminder-form edit=false}}
